minor util fixes
