+++ pkg/util/util.go
@@ -1,3 +1,17 @@
+// Copyright 2019 Istio Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package util
 
 import (
@@ -5,14 +19,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/rand"
-	"reflect"
 	"strings"
 	"time"
 
 	"github.com/ghodss/yaml"
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/protobuf/jsonpb"
-	"github.com/kr/pretty"
 	"github.com/kylelemons/godebug/diff"
 )
 
@@ -27,6 +39,18 @@ var (
 	letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
 )
 
+// Tree is a tree.
+type Tree map[string]interface{}
+
+// String implements the Stringer interface method.
+func (t Tree) String() string {
+	y, err := yaml.Marshal(t)
+	if err != nil {
+		return err.Error()
+	}
+	return string(y)
+}
+
 // RandomString returns a random string of length n.
 func RandomString(n int) string {
 	b := make([]rune, n)
@@ -36,6 +60,7 @@ func RandomString(n int) string {
 	return string(b)
 }
 
+// PrettyJSON returns the input JSON properly indented.
 func PrettyJSON(b []byte) []byte {
 	var out bytes.Buffer
 	err := json.Indent(&out, b, "", "  ")
@@ -45,157 +70,7 @@ func PrettyJSON(b []byte) []byte {
 	return out.Bytes()
 }
 
-// IsValueNil returns true if either value is nil, or has dynamic type {ptr,
-// map, slice} with value nil.
-func IsValueNil(value interface{}) bool {
-	if value == nil {
-		return true
-	}
-	switch reflect.TypeOf(value).Kind() {
-	case reflect.Slice, reflect.Ptr, reflect.Map:
-		return reflect.ValueOf(value).IsNil()
-	}
-	return false
-}
-
-// IsEmptyString returns true if value is an empty string.
-func IsEmptyString(value interface{}) bool {
-	if value == nil {
-		return true
-	}
-	switch reflect.TypeOf(value).Kind() {
-	case reflect.String:
-		return value.(string) == ""
-	}
-	return false
-}
-
-func IsString(value interface{}) bool {
-	return reflect.TypeOf(value).Kind() == reflect.String
-}
-
-// IsSlice reports whether value is a slice type.
-func IsSlice(value interface{}) bool {
-	return reflect.TypeOf(value).Kind() == reflect.Slice
-}
-
-// IsSlicePtr reports whether v is a slice ptr type.
-func IsSlicePtr(v interface{}) bool {
-	t := reflect.TypeOf(v)
-	return t.Kind() == reflect.Ptr && t.Elem().Kind() == reflect.Slice
-}
-
-// IsSliceInterfacePtr reports whether v is a slice ptr type.
-func IsSliceInterfacePtr(v interface{}) bool {
-	// Must use ValueOf because Elem().Elem() type resolves dynamically.
-	vv := reflect.ValueOf(v)
-	return vv.Kind() == reflect.Ptr && vv.Elem().Kind() == reflect.Interface && vv.Elem().Elem().Kind() == reflect.Slice
-}
-
-// IsPtr reports whether value is a ptr type.
-func IsPtr(value interface{}) bool {
-	return reflect.TypeOf(value).Kind() == reflect.Ptr
-}
-
-// IsInterfacePtr reports whether v is a slice ptr type.
-func IsInterfacePtr(v interface{}) bool {
-	t := reflect.TypeOf(v)
-	return t.Kind() == reflect.Ptr && t.Elem().Kind() == reflect.Interface
-}
-
-// IsMap reports whether value is a map type.
-func IsMap(value interface{}) bool {
-	return reflect.TypeOf(value).Kind() == reflect.Map
-}
-
-// IsMapPtr reports whether v is a map ptr type.
-func IsMapPtr(v interface{}) bool {
-	t := reflect.TypeOf(v)
-	return t.Kind() == reflect.Ptr && t.Elem().Kind() == reflect.Map
-}
-
-// IsNilOrInvalidValue reports whether v is nil or reflect.Zero.
-func IsNilOrInvalidValue(v reflect.Value) bool {
-	return !v.IsValid() || (v.Kind() == reflect.Ptr && v.IsNil()) || IsValueNil(v.Interface())
-}
-
-// AppendToSlicePtr inserts value into parent which must be a slice ptr.
-func AppendToSlicePtr(parentSlice interface{}, value interface{}) error {
-	dbgPrint("AppendToSlicePtr slice=\n%s\nvalue=\n%v", pretty.Sprint(parentSlice), value)
-	pv := reflect.ValueOf(parentSlice)
-	v := reflect.ValueOf(value)
-
-	if !IsSliceInterfacePtr(parentSlice) {
-		return fmt.Errorf("AppendToSlicePtr parent type is %T, must be *[]interface{}", parentSlice)
-	}
-
-	pv.Elem().Set(reflect.Append(pv.Elem(), v))
-
-	return nil
-}
-
-func DeleteFromSlicePtr(parentSlice interface{}, index int) error {
-	dbgPrint("DeleteFromSlicePtr index=%d, slice=\n%s", index, pretty.Sprint(parentSlice))
-	pv := reflect.ValueOf(parentSlice)
-
-	if !IsSliceInterfacePtr(parentSlice) {
-		return fmt.Errorf("DeleteFromSlicePtr parent type is %T, must be *[]interface{}", parentSlice)
-	}
-
-	pvv := pv.Elem()
-	if pvv.Kind() == reflect.Interface {
-		pvv = pvv.Elem()
-	}
-
-	ns := reflect.AppendSlice(pvv.Slice(0, index), pvv.Slice(index+1, pvv.Len()))
-	pv.Elem().Set(ns)
-
-	return nil
-}
-
-func UpdateSlicePtr(parentSlice interface{}, index int, value interface{}) error {
-	dbgPrint("UpdateSlicePtr parent=\n%s\n, index=%d, value=\n%v", pretty.Sprint(parentSlice), index, value)
-	pv := reflect.ValueOf(parentSlice)
-	v := reflect.ValueOf(value)
-
-	if !IsSliceInterfacePtr(parentSlice) {
-		return fmt.Errorf("UpdateSlicePtr parent type is %T, must be *[]interface{}", parentSlice)
-	}
-
-	pvv := pv.Elem()
-	if pvv.Kind() == reflect.Interface {
-		pv.Elem().Elem().Index(index).Set(v)
-		return nil
-	}
-	pv.Elem().Index(index).Set(v)
-
-	return nil
-}
-
-// InsertIntoMap inserts value with key into parent which must be a map, map ptr, or interface to map.
-func InsertIntoMap(parentMap interface{}, key interface{}, value interface{}) error {
-	dbgPrint("InsertIntoMap key=%v, value=%s, map=\n%s", key, pretty.Sprint(value), pretty.Sprint(parentMap))
-	v := reflect.ValueOf(parentMap)
-	kv := reflect.ValueOf(key)
-	vv := reflect.ValueOf(value)
-
-	if v.Type().Kind() == reflect.Ptr {
-		v = v.Elem()
-	}
-	if v.Type().Kind() == reflect.Interface {
-		v = v.Elem()
-	}
-
-	if v.Type().Kind() != reflect.Map {
-		dbgPrint("error %v", v.Type().Kind())
-		return fmt.Errorf("InsertIntoMap parent type is %T, must be map", parentMap)
-	}
-
-	v.SetMapIndex(kv, vv)
-
-	return nil
-}
-
+// IsYAMLEqual compares two YAML string representations by converting them to JSON first.
 func IsYAMLEqual(a, b string) bool {
 	if strings.TrimSpace(a) == "" && strings.TrimSpace(b) == "" {
 		return true
@@ -214,6 +89,7 @@ func IsYAMLEqual(a, b string) bool {
 	return string(ajb) == string(bjb)
 }
 
+// YAMLDiff returns the diff between two YAML string representations.
 func YAMLDiff(a, b string) string {
 	ao, bo := make(map[string]interface{}), make(map[string]interface{})
 	if err := yaml.Unmarshal([]byte(a), &ao); err != nil {
